The "multisearch-with-files-matching:" token now works as can be expected (which isn`t necessarily what one does expect at the first glance).
Example:

// multisearch-with-subdirs-of: "../../include/"
// multisearch-with-subdirs-of: "../"
// multisearch-with-subdirs-of: "../../lib"
// multisearch-with-files-matching: "\.\(cpp\|h\)$"

This will include ALL files *.cpp and *.h in ALL the specified
directories (in addition to files specified otherwise).  To include
files in the current directory, use

// multisearch-with-dir: "."

unless files in the current directory are otherwise included, like in
above example.

The regexp given in "multisearch-with-files-matching:" is what one
would use with `re-search-forward'.

Delimiters like double-quotes or "<" are now REQUIRED to mark the
beginning of the keys of the key-value pairs.  When the key ends at
the end of the line, the delimiter at the end of the line can be
omitted.  Missing delimiters may yield a "syntax error" message.

When you`re inclined to experiment, call

(multisearch-make-directory-files-list "example.c")

in ielm.